Cambios

+++ Funciones.py
@@ -20,9 +20,9 @@ def distanceP(velocity, density, protons, neutrons, particula):
 #Distancia esfera
 def distanceS(radio, protons, neutrons, velocity, particula, cargaEsfera):
     if particula == "Positrón":
-        return abs((2*math.pi*(8.85*10**-12)*float(radio)**2*(int(protons)+int(neutrons))*(9.11*10**-31)*int(velocity)**2)/(int(protons)*(1.6*10**-19)))
+        return abs((2*math.pi*(8.85*10**-12)*float(radio)**2*(float(protons)+float(neutrons))*(9.11*10**-31)*float(velocity)**2)/(float(protons)*(1.6*10**-19)))
     else :
-        return abs((2*math.pi*(8.85*10**-12)*float(radio)**2*(int(protons)+int(neutrons))*(1.67*10**-27)*int(velocity)**2)/(int(protons)*(1.6*10**-19)))
+        return abs((2*math.pi*(8.85*10**-12)*float(radio)**2*(float(protons)+float(neutrons))*(1.67*10**-27)*float(velocity)**2)/(float(protons)*(1.6*10**-19)))
     
 #Distancia plano infinito particula personalizada
 def distancePP(velocity, density, masa, carga):
@@ -30,7 +30,7 @@ def distancePP(velocity, density, masa, carga):
 
 #Distancia esfera particula personalizada
 def distanceSP(radio, masa, velocity, carga):
-    return abs((2*math.pi*(8.85*10**-12)*float(radio)**2*float(masa)*int(velocity)**2)/float(carga))
+    return abs((2*math.pi*(8.85*10**-12)*float(radio)**2*float(masa)*float(velocity)**2)/float(carga))
 
 #Detrminación de decimales
 def decimals(numero):
